Case 27689; adstract checkout class to reduce code duplication
