+++ src/Command/SiteCheckoutBranchCommand.php
@@ -19,22 +19,7 @@ use DennisDigital\Drupal\Console\Exception\SiteCommandException;
  *
  * @package DennisDigital\Drupal\Console\Command
  */
-class SiteCheckoutBranchCommand extends SiteBaseCommand {
-
-  /**
-   * Stores repo information.
-   *
-   * @var array Repo.
-   */
-  protected $repo;
-
-  /**
-   * Stores branch information.
-   *
-   * @var array Branch.
-   */
-  protected $branch;
-
+class SiteCheckoutBranchCommand extends AbstractSiteCheckoutCommand {
   /**
    * Stores current branch of the checked out code.
    *
@@ -53,16 +38,11 @@ class SiteCheckoutBranchCommand extends SiteBaseCommand {
 
     // Custom options.
     $this->addOption(
-        'ignore-changes',
-        '',
-        InputOption::VALUE_NONE,
-        'Ignore local changes when checking out the site'
-      )->addOption(
-        'branch',
-        '-B',
-        InputOption::VALUE_OPTIONAL,
-        'Specify which branch to checkout if different than the global branch found in sites.yml'
-      );
+      'branch',
+      '-B',
+      InputOption::VALUE_OPTIONAL,
+      'Specify which branch to checkout if different than the global branch found in sites.yml'
+    );
   }
 
   /**
@@ -70,9 +50,6 @@ class SiteCheckoutBranchCommand extends SiteBaseCommand {
    */
   protected function interact(InputInterface $input, OutputInterface $output) {
     parent::interact($input, $output);
-
-    // Validate repo.
-    $this->validateRepo();
 
     $remoteBranches = $this->getRemoteBranches();
     $defaultBranch = $this->getDefaultBranch();
@@ -105,20 +82,17 @@ class SiteCheckoutBranchCommand extends SiteBaseCommand {
   protected function execute(InputInterface $input, OutputInterface $output) {
     parent::execute($input, $output);
 
-    // Validate repo.
-    $this->validateRepo();
-
     // Validate branch.
     $this->validateBranch($input);
 
-    if ($this->branch == $this->currentBranch) {
+    if ($this->ref == $this->currentBranch) {
       $this->io->commentBlock('Current branch selected, skipping checkout command.');
       return;
     }
 
     $this->io->comment(sprintf('Checking out %s (%s) on %s',
       $this->siteName,
-      $this->branch,
+      $this->ref,
       $this->destination
     ));
 
@@ -152,24 +126,6 @@ class SiteCheckoutBranchCommand extends SiteBaseCommand {
   }
 
   /**
-   * Helper to validate repo.
-   *
-   * @throws SiteCommandException
-   *
-   * @return string Repo url
-   */
-  protected function validateRepo() {
-    if (isset($this->config['repo'])) {
-      $this->repo = $this->config['repo'];
-    }
-    else {
-      throw new SiteCommandException('Repo not found in sites.yml');
-    }
-
-    return $this->repo;
-  }
-
-  /**
    * Helper to validate branch option.
    *
    * @param InputInterface $input
@@ -181,110 +137,20 @@ class SiteCheckoutBranchCommand extends SiteBaseCommand {
       !is_null($input->getOption('branch'))
     ) {
       // Use config from parameter.
-      $this->branch = $input->getOption('branch');
+      $this->ref = $input->getOption('branch');
     }
     elseif (isset($this->config['repo']['branch'])) {
       // Use config from sites.yml.
-      $this->branch = $this->config['repo']['branch'];
+      $this->ref = $this->config['repo']['branch'];
     }
     else {
-      $this->branch = '8.x';
+      $this->ref = '8.x';
     }
 
     // Update input.
     if ($input->hasOption('branch')) {
-      $input->setOption('branch', $this->branch);
+      $input->setOption('branch', $this->ref);
     }
-  }
-
-  /**
-   * Helper to detect local modifications.
-   *
-   * @return TRUE If everything is ok.
-   *
-   * @throws SiteCommandException
-   */
-  protected function gitDiff() {
-    $command = sprintf(
-      'cd %s && git diff-files --name-status -r --ignore-submodules',
-      $this->shellPath($this->destination)
-    );
-
-    $shellProcess = $this->getShellProcess();
-
-    if ($shellProcess->exec($command, TRUE)) {
-      if (!empty($shellProcess->getOutput())) {
-        $message = sprintf('You have uncommitted changes on %s' . PHP_EOL .
-          'Please commit or revert your changes before checking out the site.' . PHP_EOL .
-          'If you want to check out the site without committing the changes use --ignore-changes.',
-          $this->destination
-        );
-        throw new SiteCommandException($message);
-      }
-    }
-    else {
-      throw new SiteCommandException($shellProcess->getOutput());
-    }
-
-    return TRUE;
-  }
-
-  /**
-   * Helper to do the actual clone command.
-   *
-   * @return bool TRUE If successful.
-   *
-   * @throws SiteCommandException
-   */
-  protected function gitClone() {
-    $command = sprintf('git clone --branch %s %s %s',
-      $this->branch,
-      $this->repo['url'],
-      $this->shellPath($this->destination)
-    );
-    $this->io->commentBlock($command);
-
-    $shellProcess = $this->getShellProcess();
-
-    if ($shellProcess->exec($command, TRUE)) {
-      $this->io->success(sprintf('Repo cloned on %s', $this->destination));
-    }
-    else {
-      throw new SiteCommandException($shellProcess->getOutput());
-    }
-
-    return TRUE;
-  }
-
-  /**
-   * Helper to check out a branch.
-   *
-   * @return bool TRUE If successful.
-   *
-   * @throws SiteCommandException
-   */
-  protected function gitCheckout() {
-    $command = sprintf(
-        'cd %s && ' .
-        'git fetch --all && ' .
-        'chmod 777 web/sites/default && ' .
-        'chmod 777 web/sites/default/settings.php && ' .
-        'git checkout %s ',
-      $this->shellPath($this->destination),
-      $this->branch
-    );
-
-    $shellProcess = $this->getShellProcess();
-
-    if ($shellProcess->exec($command, TRUE)) {
-      $this->io->success(sprintf('Checked out %s', $this->branch));
-    }
-    else {
-      throw new SiteCommandException($shellProcess->getOutput());
-
-    }
-
-    return TRUE;
   }
 
   /**
