+++ src/Command/SiteCheckoutTagCommand.php
@@ -19,22 +19,7 @@ use DennisDigital\Drupal\Console\Exception\SiteCommandException;
  *
  * @package DennisDigital\Drupal\Console\Command
  */
-class SiteCheckoutTagCommand extends SiteBaseCommand {
-
-  /**
-   * Stores repo information.
-   *
-   * @var array Repo.
-   */
-  protected $repo;
-
-  /**
-   * Stores tag information.
-   *
-   * @var string tag.
-   */
-  protected $tag;
-
+class SiteCheckoutTagCommand extends AbstractSiteCheckoutCommand {
   /**
    * Stores current tag of the checked out code.
    *
@@ -53,16 +38,11 @@ class SiteCheckoutTagCommand extends SiteBaseCommand {
 
     // Custom options.
     $this->addOption(
-        'ignore-changes',
-        '',
-        InputOption::VALUE_NONE,
-        'Ignore local changes when checking out the site'
-      )->addOption(
-        'tag',
-        '-T',
-        InputOption::VALUE_OPTIONAL,
-        'Specify which tag to checkout if different than the global tag found in sites.yml'
-      );
+      'tag',
+      '-T',
+      InputOption::VALUE_OPTIONAL,
+      'Specify which tag to checkout if different than the global tag found in sites.yml'
+    );
   }
 
   /**
@@ -70,9 +50,6 @@ class SiteCheckoutTagCommand extends SiteBaseCommand {
    */
   protected function interact(InputInterface $input, OutputInterface $output) {
     parent::interact($input, $output);
-
-    // Validate repo.
-    $this->validateRepo();
 
     $tag = $input->getOption('tag');
     if (!$tag) {
@@ -90,20 +67,17 @@ class SiteCheckoutTagCommand extends SiteBaseCommand {
   protected function execute(InputInterface $input, OutputInterface $output) {
     parent::execute($input, $output);
 
-    // Validate repo.
-    $this->validateRepo();
-
     // Validate tag.
     $this->validateTag($input);
 
-    if ($this->tag == $this->currentTag) {
+    if ($this->ref == $this->currentTag) {
       $this->io->commentBlock('Current tag selected, skipping checkout command.');
       return;
     }
 
     $this->io->comment(sprintf('Checking out %s (%s) on %s',
       $this->siteName,
-      $this->tag,
+      $this->ref,
       $this->destination
     ));
 
@@ -137,24 +111,6 @@ class SiteCheckoutTagCommand extends SiteBaseCommand {
   }
 
   /**
-   * Helper to validate repo.
-   *
-   * @throws SiteCommandException
-   *
-   * @return string Repo url
-   */
-  protected function validateRepo() {
-    if (isset($this->config['repo'])) {
-      $this->repo = $this->config['repo'];
-    }
-    else {
-      throw new SiteCommandException('Repo not found in sites.yml');
-    }
-
-    return $this->repo;
-  }
-
-  /**
    * Helper to validate tag option.
    *
    * @param InputInterface $input
@@ -166,103 +122,11 @@ class SiteCheckoutTagCommand extends SiteBaseCommand {
       !is_null($input->getOption('tag'))
     ) {
       // Use config from parameter.
-      $this->tag = $input->getOption('tag');
+      $this->ref = $input->getOption('tag');
     }
     else {
       throw new SiteCommandException('Tag must be specified.');
     }
-  }
-
-  /**
-   * Helper to detect local modifications.
-   *
-   * @return TRUE If everything is ok.
-   *
-   * @throws SiteCommandException
-   */
-  protected function gitDiff() {
-    $command = sprintf(
-      'cd %s && git diff-files --name-status -r --ignore-submodules',
-      $this->shellPath($this->destination)
-    );
-
-    $shellProcess = $this->getShellProcess();
-
-    if ($shellProcess->exec($command, TRUE)) {
-      if (!empty($shellProcess->getOutput())) {
-        $message = sprintf('You have uncommitted changes on %s' . PHP_EOL .
-          'Please commit or revert your changes before checking out the site.' . PHP_EOL .
-          'If you want to check out the site without committing the changes use --ignore-changes.',
-          $this->destination
-        );
-        throw new SiteCommandException($message);
-      }
-    }
-    else {
-      throw new SiteCommandException($shellProcess->getOutput());
-    }
-
-    return TRUE;
-  }
-
-  /**
-   * Helper to do the actual clone command.
-   *
-   * @return bool TRUE If successful.
-   *
-   * @throws SiteCommandException
-   */
-  protected function gitClone() {
-    $command = sprintf('git clone %s %s',
-      $this->repo['url'],
-      $this->shellPath($this->destination)
-    );
-    $this->io->commentBlock($command);
-
-    $shellProcess = $this->getShellProcess();
-
-    if ($shellProcess->exec($command, TRUE)) {
-      $this->io->success(sprintf('Repo cloned on %s', $this->destination));
-    }
-    else {
-      throw new SiteCommandException($shellProcess->getOutput());
-    }
-
-    // Checkout the tag.
-    $this->gitCheckout();
-
-    return TRUE;
-  }
-
-  /**
-   * Helper to check out a tag.
-   *
-   * @return bool TRUE If successful.
-   *
-   * @throws SiteCommandException
-   */
-  protected function gitCheckout() {
-    $command = sprintf(
-        'cd %s && ' .
-        'git fetch --all && ' .
-        'chmod 777 web/sites/default && ' .
-        'chmod 777 web/sites/default/settings.php && ' .
-        'git checkout %s ',
-      $this->shellPath($this->destination),
-      $this->tag
-    );
-
-    $shellProcess = $this->getShellProcess();
-
-    if ($shellProcess->exec($command, TRUE)) {
-      $this->io->success(sprintf('Checked out %s', $this->tag));
-    }
-    else {
-      throw new SiteCommandException($shellProcess->getOutput());
-
-    }
-
-    return TRUE;
   }
 
 }
